rename: gateway 이름 변경
